Add MermaidCycleDetector
