+++ src/view/output-formatter.ts
@@ -1,12 +1,14 @@
 import * as vscode from "vscode";
 import { MermaidLinkGenerator } from "./mermaid/link-generator";
 import { telemetry } from "../telemetry";
+import { MermaidCycleDetector } from "./mermaid/cycle-detector";
 
 export class OutputFormatter {
     public static getDiagramFileContent(modelName: string, llmResponse: string): string {
         const mermaidBlock = this.getMermaidBlock(llmResponse);
         const mermaidCode = mermaidBlock.replace(/```mermaid|```/g, "");
-        this.tryDetectCycle(mermaidCode);
+        const cycleDetector = new MermaidCycleDetector(mermaidCode);
+        cycleDetector.detectCycle();
         const linkGenerator = new MermaidLinkGenerator(mermaidCode);
 
         return `<p align="center">
@@ -41,51 +43,6 @@ ${mermaidBlock}`;
         }
 
         return block;
-    }
-
-    public static tryDetectCycle(mermaidCode: string): void {
-        try {
-            const cycleNode = this.detectCycle(mermaidCode);
-
-            if (cycleNode) {
-                console.log(`Cycle detected in the diagram at node: ${cycleNode}`);
-                telemetry.sendTelemetryErrorEvent("diagramCycleDetected");
-            }
-        } catch (error) {
-            telemetry.sendTelemetryErrorEvent("diagramCycleDetectionFailed");
-        }
-    }
-
-    public static detectCycle(mermaidCode: string): string | undefined {
-        const lines = mermaidCode.split("\n");
-        const parentNodes: string[] = [];
-
-        for (let line of lines) {
-            line = line.trim();
-
-            if (line.startsWith("subgraph")) {
-                const rest = line.substring("subgraph".length, line.length);
-                const subgraphName = rest.split("[")[0].trim();
-
-                if (parentNodes.includes(subgraphName)) {
-                    return subgraphName;
-                }
-
-                parentNodes.push(subgraphName);
-            } else if (line.startsWith("end")) {
-                parentNodes.pop();
-            } else if (line === "") {
-                continue;
-            } else {
-                const node = line.split("[")[0];
-
-                if (parentNodes.includes(node)) {
-                    return node;
-                }
-            }
-        }
-
-        return undefined;
     }
 
     public static getLogFileContent(
